requirements.txt
pandas>=2.0.0
neo4j>=5.0.0
openpyxl>=3.0.0  # Required for pandas Excel support
python-dotenv>=1.0.0  # For environment variables management

# RAG (Retrieval Augmented Generation) dependencies
langchain>=0.0.350
langchain-community>=0.0.10
langchain-ollama>=0.1.0  # Updated Ollama integration
langchain-neo4j>=0.1.0  # Updated Neo4j integration
ollama>=0.1.0

# Fast Hybrid RAG dependencies
sentence-transformers>=2.2.2  # For fast embeddings
scikit-learn>=1.3.0  # For cosine similarity calculations

# Web application dependencies
flask>=2.3.0  # Minimal web framework